error correction in case the user is located eastbound of UTC
